Revert "Start AnnoyerService in SmsReceiver"
This reverts commit 60f0b1f7e1698bf3dcb8a047fdc50c27171affa4.

Starting in API 26, the system imposes tighter restrictions on what an
app may do in the background: if the app itself is not in the
foreground (by e.g. having an activity visible on the screen), it may
not run a background service. So, if we launch the AnnoyerService at the
same time as the SmsReceiver, and the service starts first, it will be
killed. (We can avoid this by making the service a foreground service,
displaying a notification that the app is running; but then if something
goes wrong, the service can't be killed by swiping away the app in the
task manager. Maybe ultimately this is preferable behaviour?)

Launching the annoyer from the sms receiver also means that the "test
alert UI" button doesn't (automatically) trigger the annoyer.

Maybe a foreground service is indeed the best approach; this needs more
thought.